+++ src/mkvtoolnix-gui/main_window/preferences_dialog.cpp
@@ -221,6 +221,11 @@ PreferencesDialog::setupToolTips() {
                    .arg(QY("If enabled, automatic scaling for high DPI displays will be disabled."))
                    .arg(QY("Changes to this option will only take effect the next time the application is started.")));
 
+  Util::setToolTip(ui->cbGuiDisableDarkStyleSheet,
+                   Q("%1 %2")
+                   .arg(QY("By default the GUI will start up with a dark color scheme if the color scheme in Windows is set to dark mode."))
+                   .arg(QY("If this option is enabled, the GUI will not use its dark mode and fall back to the default color scheme.")));
+
   Util::setToolTip(ui->cbGuiCheckForUpdates,
                    Q("%1 %2 %3")
                    .arg(QY("If enabled, the program will check online whether or not a new release of MKVToolNix is available on the home page."))
@@ -767,6 +772,9 @@ PreferencesDialog::setupFontAndScaling() {
   ui->sbGuiFontPointSize->setValue(font.pointSize());
 
   ui->cbGuiDisableHighDPIScaling->setChecked(m_cfg.m_uiDisableHighDPIScaling);
+  ui->cbGuiDisableDarkStyleSheet->setChecked(m_cfg.m_uiDisableDarkStyleSheet);
+#if !defined(SYS_WINDOWS)
+  ui->cbGuiDisableDarkStyleSheet->setVisible(false);
 #endif
 }
 
@@ -778,6 +786,7 @@ PreferencesDialog::save() {
   m_cfg.m_uiFontFamily                                  = ui->fcbGuiFontFamily->currentFont().family();
   m_cfg.m_uiFontPointSize                               = ui->sbGuiFontPointSize->value();
   m_cfg.m_uiDisableHighDPIScaling                       = ui->cbGuiDisableHighDPIScaling->isChecked();
+  m_cfg.m_uiDisableDarkStyleSheet                       = ui->cbGuiDisableDarkStyleSheet->isChecked();
   m_cfg.m_checkForUpdates                               = ui->cbGuiCheckForUpdates->isChecked();
   m_cfg.m_showToolSelector                              = ui->cbGuiShowToolSelector->isChecked();
   m_cfg.m_warnBeforeClosingModifiedTabs                 = ui->cbGuiWarnBeforeClosingModifiedTabs->isChecked();