+++ src/mkvtoolnix-gui/util/settings.h
@@ -176,7 +176,7 @@ public:
   bool m_useDefaultJobDescription, m_showOutputOfAllJobs, m_switchToJobOutputAfterStarting, m_resetJobWarningErrorCountersOnExit;
   bool m_removeOutputFileOnJobFailure;
 
-  bool m_uiDisableHighDPIScaling;
+  bool m_uiDisableHighDPIScaling, m_uiDisableDarkStyleSheet;
   bool m_checkForUpdates;
   QDateTime m_lastUpdateCheck;